add control signals for HDMI preamble
This adds output ports for the HDMI preamble signals to the IP and
allows the IP to be used as a real HDMI sink. The HDMI protocol
slightly differs from DVI in the sense that during the blank phases
so-called data islands can be transferred (mostly for audio). Those
data islands are announced via the HDMI preamble. The preamble is
transferred from the source to the sink via the C0 und C1 control
signals of the TMDS channels 1 and 2 (green and red). The 4 preamble
bits are therfore transferred in exactly the same manner as the vsync
and hsync signals (which are send through the blue channel).
Without the preamble signal it is not possible to distinguish between
data islands and video data.
Please note that the IP in this version will still identify as a
DVI sink by the EDIDs that are packed with the IP. The changes of this
commit were made to allow the usage of the IP as a sink to connect to
an HDMI camera which ignores the EDID (at least the video standard part)
and operates in HDMI mode anyway. If you want to provoke any source
to operate in HDMI mode the EDIDs would have to be changed. (See
HDMI standard 1.3 Appendix C on details).