fix reversed channels for rgb2dvi an dvi2rgb ip-cores
For the two IP cores the order of the channels was reversed
in both IP cores. Blue was mapped to bits 15:8 and green to 7:0
instead of the standard RGB ordering.
The code for both cores cotained a comment that "for some reason"
the order is reversed. Potentially this was tested with a video
source where the channels were reversed or on a board were the
HDMI/DVI pins for blue and green were reserved.
However, at least on the Nexys Video, the video data is delivered
as expected from the deserializers and there is no need to reverse
the channels. I.e Blue on channel 0, green on channel 1 and red
on channel 2. The assignment of the sync signals was always correct
(channel 0) since they are stripped before the mapping on
the 24 bit data vector happens.